# BaaSPlatform/contact-service/src/main/resources/META-INF/resources/api.yml
openapi: 3.0.1
info:
  title: contact-service API
  version: 1.0.0
  description: |
    Contact Service
servers:
  - url: http://localhost:8074
    description: Generated server url
...
paths:
...
  /beneficiaries:
    get:
      tags:
        - Beneficiary
      summary: Lookup beneficiary account
      description: Lookup beneficiary account by mobile number and account number
      operationId:  lookupBeneficiary
      parameters:
        - name: mobileNumber
          in: query
          description: Mobile Number
          required: false
          schema:
            type: string
          example: 65086508427
        - name: accountNumber
          in: query
          description: Account Number
          required: false
          schema:
            type: string
          example: 120086508427
        - name: accountType
          in: query
          required: false
          schema:
            type: string
          example: BANK_WALLET
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
    post:
      tags:
        - Beneficiary
      summary: Register NAD
      operationId: createBeneficiary
      requestBody:
        description: Beneficiary data model
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBeneficiaryRecord'
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryCreatedData'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
  /beneficiaries/{type}/{identification}:
    put:
      tags:
        - Beneficiary
      summary: Update DuitNow Id
      operationId: updateBeneficiary
      parameters:
        - name: type
          in: path
          description: beneficiary type
          required: true
          schema:
            type: string
          example: MobileNumber
        - name: identification
          in: path
          description: identification value
          required: true
          schema:
            type: string
          example: 609100000147
      requestBody:
        description: Beneficiary data model
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBeneficiaryRecord'
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryAccountDetails'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
    get: 
      tags:
        - Beneficiary
      summary: Lookup Beneficiary By Identification
      operationId: lookupBeneficiaryAccount
      parameters:
        - name: type
          in: path
          description: beneficiary type
          required: true
          schema:
            type: string
          example: MobileNumber
        - name: identification
          in: path
          description: identification value
          required: true
          schema:
            type: string
          example: 609100000147
        - name: bankCode
          in: query
          required: false
          schema:
            type: string
          example: RJHIMYKL
        - name: walletId
          in: query
          required: false
          schema:
            type: string
          example: f3b9595e-80fc-4333-9611-b0fe4a3ef222
        - name: beneficiaryAccountType
          in: query
          required: false
          schema:
            type: string
            enum:
              - SavingsAccount
              - CurrentAccount
              - LoanAccount
              - CreditCard
              - CASA
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupBeneficiaryResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Beneficiary
      summary: Delete Beneficiary By Identification
      operationId: deleteBeneficiary
      parameters:
        - name: type
          in: path
          description: beneficiary type
          required: true
          schema:
            type: string
          example: MobileNumber
        - name: identification
          in: path
          description: identification value
          required: true
          schema:
            type: string
          example: 609100000147
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Deleted succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryDeleteResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
  /beneficiaries/QR/validations:
    post:
      tags:
        - Beneficiary
      summary: Lookup Beneficiary by QR code
      operationId: lookupBeneficiaryByQrCode
      requestBody:
        description: Lookup Beneficiary QR data model
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeneficiaryValidation'
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupBeneficiaryResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]        
#-----
components:
  schemas:
    LookupBeneficiaryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Beneficiary'
...
    PagingInformation:
      type: object
      properties:
        totalRecords:
          type: integer
        pageSize:
          type: integer
        pageNumber:
          type: integer
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ResponseStatus'
    ResponseStatus:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    WalletResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Beneficiary'
        paging:
          $ref: '#/components/schemas/PagingInformation'
        status:
          $ref: '#/components/schemas/ResponseStatus'
...
    Beneficiary:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        paymentReference:
          type: string
        accountNumber:
          type: string
        bankCode:
          type: string
        accountType:
          type: string
        status:
          type: string
        identification:
          type: string
        type:
            type: string
        qrDetails:
          readOnly: true
          $ref: '#/components/schemas/QrDetails'            
    QrDetails:
      type: object
      properties:
        qrType:
          type: string
          enum:
            - static
            - dynamic
        amount:
          type: string
        messageId:
          type: string
        acquirerId:
          type: string
        acquirerName:
          type: string
        category:
          type: string
          enum:
            - POS
            - P2P
            - DFLT
        acceptedSource:
          type: array
          items:
            type: string
        merchantCode:
          type: string
        qrRawData:
          type: string
        originalMessageId:
          type: string 
          example: 20230301FNXSMYNB520HQR02000043 
    BeneficiaryValidation:
      type: object
      properties:
        qrRawData:
          type: string
        walletId:
          type: string           
...
    UpdateBeneficiaryRecord:
      type: object
      properties:
        secondaryType:
          type: string
          example: PassportNumber
        secondaryIdentification:
          type: string
          example: E394039340VSGP
        displayName:
          type: string
          example: Test Name
        paymentReference:
          type: string
          example: 0011002932913
        bankCode:
          type: string
          example: FNXSMYNB
        status:
          type: string
          enum:
            - Inactive
            - Active
            - Deactivate
            - BankSuspended
    BeneficiaryDataResponse:
      type: object
      properties:
        paymentReference:
          type: string
        displayName:
          type: string
        bankCode:
          type: string
        identification:
          type: string
        type:
          type: string
        secondaryType:
          type: string
        secondaryIdentification:
          type: string
        status:
          type: string
...
    BeneficiaryAccountDetails:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BeneficiaryDataResponse'
...
    CreateBeneficiaryRecord:
      type: object
      properties:
        type:
          type: string
        identification:
          type: string
        secondaryType:
          type: string
        secondaryIdentification:
          type: string
        displayName:
          type: string
          default: ""
          x-constraints: "@Xss"
        paymentReference:
          type: string
          default: ""
          x-constraints: "@Xss"
        bankCode:
          type: string
          default: ""
...
    BeneficiaryCreatedData:
      type: object
      properties:
        paymentReference:
          type: string
        displayName:
          type: string
        bankCode:
          type: string
        identification:
          type: string
        type:
          type: string
        secondaryType:
          type: string
        secondaryIdentification:
          type: string
        status:
          type: string
...
    BeneficiaryDeleteResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BeneficiaryDeleteData'
    BeneficiaryDeleteData:
      type: object
      properties:
        paymentReference:
          type: string
        displayName:
          type: string
        bankCode:
          type: string
        identification:
          type: string
        type:
          type: string
        secondaryType:
          type: string
        secondaryIdentification:
          type: string
        status:
          type: string
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer